# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/admin/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin: Authentication"
                ],
                "summary": "Administrator Login",
                "parameters": [
                    {
                        "description": "JSON request body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swagger.adminLoginBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminLogin200"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminLogin401"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminLogin403"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminLogin422"
                        }
                    }
                }
            }
        },
        "/admin/logout": {
            "post": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin: Authentication"
                ],
                "summary": "Administrator Logout",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminLogout200"
                        }
                    }
                }
            }
        },
        "/admin/refresh": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin: Authentication"
                ],
                "summary": "Administrator Refresh Token",
                "parameters": [
                    {
                        "description": "JSON request body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swagger.adminRefreshBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminRefresh200"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminRefresh401"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminRefresh403"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminRefresh422"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "swagger.adminLogin200": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE1OTY3NjI3MTcsInJvbGUiOiJhZG1pbiIsInN1YiI6MX0.BepVrGACE5xvkA08cMDptPR0sz5fcKkPfc4NO-oxZRE"
                },
                "message": {
                    "type": "string",
                    "example": "Authenticated"
                },
                "refresh_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2MjgyOTUxMTcsInN1YiI6MX0.uS2sJ52TTe40JSvqwGTigbig186sDmXRl9FB8OeheWw"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "swagger.adminLogin401": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Please provide valid credentials"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminLogin403": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Account has been suspended"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminLogin422": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Required fields either missing or empty"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminLoginBody": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "test@test.com"
                },
                "password": {
                    "type": "string",
                    "example": "password"
                }
            }
        },
        "swagger.adminLogout200": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Logged out"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "swagger.adminRefresh200": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE1OTY3NjI3MTcsInJvbGUiOiJhZG1pbiIsInN1YiI6MX0.BepVrGACE5xvkA08cMDptPR0sz5fcKkPfc4NO-oxZRE"
                },
                "message": {
                    "type": "string",
                    "example": "Token refreshed"
                },
                "refresh_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2MjgyOTUxMTcsInN1YiI6MX0.uS2sJ52TTe40JSvqwGTigbig186sDmXRl9FB8OeheWw"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "swagger.adminRefresh401": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Invalid token"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminRefresh403": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Account has been suspended"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminRefresh422": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Required fields either missing or empty"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminRefreshBody": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2MjgzNjAyOTIsInN1YiI6MX0.5ddhUgEXlnN3NZH0yrHH1uGspHDFnqevA18BVy5dFDY"
                }
            }
        }
    },
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}